requirements.txt
# Core Framework
aiogram==3.13.1
fastapi==0.109.0
uvicorn[standard]==0.27.0
asyncpg==0.29.0
sqlalchemy==2.0.25
alembic==1.13.1
pydantic==2.5.3
pydantic-settings==2.1.0

# Security & Cryptography
cryptography==42.0.2
PyJWT==2.8.0
argon2-cffi==23.1.0
itsdangerous==2.1.2
python-jose[cryptography]==3.3.0

# Rate Limiting & Cache
slowapi==0.1.9
redis==5.0.1
hiredis==2.3.2

# Validation & Utils
email-validator==2.1.0
python-multipart==0.0.6
aiofiles==23.2.1
httpx==0.26.0
requests==2.31.0
python-dotenv==1.0.1

# External Services
cloudinary==1.38.0
sendgrid==6.11.0

# Database
psycopg2-binary==2.9.9

# Logging & Monitoring
loguru==0.7.2
colorlog==6.8.2
sentry-sdk==1.40.0

# TON Integration
pytonlib==0.0.50
tonsdk==1.0.15

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
pytest-mock==3.12.0
faker==22.6.0

# Additional Utils
Jinja2==3.1.3
python-dateutil==2.8.2
pytz==2024.1
